Support request debug logging in the builder package
The request body is only logged if the Content-Type is printable (e.g.
application/json or text/plain) to prevent payloads that are not human
readable, like the build context tar, from getting logged out.
